# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RelAI API Documentation",
        "description": "API documentation for RelAI - Realisasi Keuangan Kementerian Koordinator Bidang Infrastruktur dan Pembangunan Kewilayahan",
        "contact": {
            "name": "RelAI Admin",
            "email": "admin@relai.go.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/anggarans/{id}": {
            "get": {
                "tags": [
                    "Anggaran"
                ],
                "summary": "Get single anggaran by ID",
                "operationId": "d216d2bf9e1a18ef03872bc76625dcb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Anggaran details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "tahun": {
                                                    "type": "integer"
                                                },
                                                "total_anggaran": {
                                                    "type": "number"
                                                },
                                                "anggaran_terpakai": {
                                                    "type": "number"
                                                },
                                                "sp2d": {
                                                    "type": "number"
                                                },
                                                "sisa_anggaran": {
                                                    "type": "number"
                                                },
                                                "keterangan": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Anggaran not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Anggaran"
                ],
                "summary": "Update anggaran data",
                "operationId": "1442e8bbce108b7c3cb542cf0cb7afe9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tahun": {
                                        "type": "integer",
                                        "example": 2025
                                    },
                                    "total_anggaran": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 2500000000
                                    },
                                    "anggaran_terpakai": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 750000000
                                    },
                                    "sp2d": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 500000000
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Updated anggaran tahun 2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Anggaran updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Anggaran updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Anggaran"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Anggaran"
                ],
                "summary": "Delete anggaran",
                "operationId": "7880cf73d13f20c914a4e20579f38033",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Anggaran deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Anggaran deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Anggaran not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/anggarans": {
            "post": {
                "tags": [
                    "Anggaran"
                ],
                "summary": "Create new anggaran",
                "operationId": "34460f2ca474a68d9e382719b97398b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tahun",
                                    "total_anggaran"
                                ],
                                "properties": {
                                    "tahun": {
                                        "type": "integer",
                                        "example": 2025
                                    },
                                    "total_anggaran": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 2000000000
                                    },
                                    "anggaran_terpakai": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 0
                                    },
                                    "sp2d": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 0
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Anggaran tahun 2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Anggaran created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Anggaran created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Anggaran"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Anggaran": {
                "title": "Anggaran",
                "description": "Anggaran model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tahun": {
                        "type": "integer",
                        "example": 2025
                    },
                    "total_anggaran": {
                        "type": "number",
                        "format": "decimal",
                        "example": 2000000000
                    },
                    "anggaran_terpakai": {
                        "type": "number",
                        "format": "decimal",
                        "example": 500000000
                    },
                    "sp2d": {
                        "type": "number",
                        "format": "decimal",
                        "example": 300000000
                    },
                    "sisa_anggaran": {
                        "type": "number",
                        "format": "decimal",
                        "example": 1500000000
                    },
                    "keterangan": {
                        "type": "string",
                        "example": "Anggaran tahun 2025"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}